Separate metadata and object record publisher
