+++ packages/twenty-server/src/engine/api/graphql/graphql-query-runner/enums/database-event-action.ts
@@ -1,3 +1,5 @@
+import { registerEnumType } from '@nestjs/graphql';
+
 export enum DatabaseEventAction {
   CREATED = 'created',
   UPDATED = 'updated',
@@ -6,3 +8,8 @@ export enum DatabaseEventAction {
   RESTORED = 'restored',
   UPSERTED = 'upserted',
 }
+
+registerEnumType(DatabaseEventAction, {
+  name: 'DatabaseEventAction',
+  description: 'Database Event Action',
+});
