+++ packages/twenty-server/src/engine/core-modules/core-engine.module.ts
@@ -65,7 +65,6 @@ import { RoleModule } from 'src/engine/metadata-modules/role/role.module';
 import { RowLevelPermissionModule } from 'src/engine/metadata-modules/row-level-permission-predicate/row-level-permission.module';
 import { SubscriptionsModule } from 'src/engine/subscriptions/subscriptions.module';
 import { TrashCleanupModule } from 'src/engine/trash-cleanup/trash-cleanup.module';
-import { MetadataEventEmitterModule } from 'src/engine/metadata-event-emitter/metadata-event-emitter.module';
 import { WorkspaceEventEmitterModule } from 'src/engine/workspace-event-emitter/workspace-event-emitter.module';
 import { ChannelSyncModule } from 'src/modules/connected-account/channel-sync/channel-sync.module';
 import { DashboardModule } from 'src/modules/dashboard/dashboard.module';
@@ -110,7 +109,6 @@ import { FileModule } from './file/file.module';
     PostgresCredentialsModule,
     WorkflowApiModule,
     WorkspaceEventEmitterModule,
-    MetadataEventEmitterModule,
     ActorModule,
     TelemetryModule,
     AdminPanelModule,
