+++ packages/twenty-server/src/engine/subscriptions/dtos/on-db-event.dto.ts
@@ -1,14 +1,9 @@
-import { Field, ObjectType, registerEnumType } from '@nestjs/graphql';
+import { Field, ObjectType } from '@nestjs/graphql';
 
 import GraphQLJSON from 'graphql-type-json';
 import { ObjectRecord } from 'twenty-shared/types';
 
 import { DatabaseEventAction } from 'src/engine/api/graphql/graphql-query-runner/enums/database-event-action';
-
-registerEnumType(DatabaseEventAction, {
-  name: 'DatabaseEventAction',
-  description: 'Database Event Action',
-});
 
 @ObjectType('OnDbEvent')
 export class OnDbEventDTO {
