+++ packages/twenty-server/src/engine/subscriptions/enums/subscription-channel.enum.ts
@@ -1,5 +1,4 @@
 export enum SubscriptionChannel {
-  DATABASE_EVENT_CHANNEL = 'DATABASE_EVENT_CHANNEL',
   LOGIC_FUNCTION_LOGS_CHANNEL = 'LOGIC_FUNCTION_LOGS_CHANNEL',
   EVENT_STREAM_CHANNEL = 'EVENT_STREAM_CHANNEL',
 }
