+++ packages/twenty-server/src/engine/subscriptions/event-stream.resolver.ts
@@ -19,10 +19,8 @@ import { WorkspaceAuthGuard } from 'src/engine/guards/workspace-auth.guard';
 import { EVENT_STREAM_TTL_MS } from 'src/engine/subscriptions/constants/event-stream-ttl.constant';
 import { AddQuerySubscriptionInput } from 'src/engine/subscriptions/dtos/add-query-subscription.input';
 import { EventSubscriptionDTO } from 'src/engine/subscriptions/dtos/event-subscription.dto';
-import { OnDbEventDTO } from 'src/engine/subscriptions/dtos/on-db-event.dto';
-import { OnDbEventInput } from 'src/engine/subscriptions/dtos/on-db-event.input';
 import { RemoveQueryFromEventStreamInput } from 'src/engine/subscriptions/dtos/remove-query-subscription.input';
-import { SubscriptionChannel } from 'src/engine/subscriptions/enums/subscription-channel.enum';
+import { EventStreamExceptionFilter } from 'src/engine/subscriptions/event-stream-exception.filter';
 import {
   EventStreamException,
   EventStreamExceptionCode,
@@ -30,56 +28,18 @@ import {
 import { EventStreamService } from 'src/engine/subscriptions/event-stream.service';
 import { SubscriptionService } from 'src/engine/subscriptions/subscription.service';
 import { type EventStreamPayload } from 'src/engine/subscriptions/types/event-stream-payload.type';
-import { wrapAsyncIteratorWithLifecycle } from 'src/engine/workspace-event-emitter/utils/wrap-async-iterator-with-lifecycle';
-import { WorkspaceEventEmitterExceptionFilter } from 'src/engine/workspace-event-emitter/workspace-event-emitter-exception.filter';
-
-import { eventStreamIdToChannelId } from './utils/get-channel-id-from-event-stream-id';
+import { eventStreamIdToChannelId } from 'src/engine/subscriptions/utils/get-channel-id-from-event-stream-id';
+import { wrapAsyncIteratorWithLifecycle } from 'src/engine/subscriptions/utils/wrap-async-iterator-with-lifecycle';
 
 @MetadataResolver()
 @UseGuards(WorkspaceAuthGuard, UserAuthGuard, NoPermissionGuard)
 @UsePipes(ResolverValidationPipe)
-@UseFilters(
-  WorkspaceEventEmitterExceptionFilter,
-  PreventNestToAutoLogGraphqlErrorsFilter,
-)
-export class WorkspaceEventEmitterResolver {
+@UseFilters(EventStreamExceptionFilter, PreventNestToAutoLogGraphqlErrorsFilter)
+export class EventStreamResolver {
   constructor(
     private readonly subscriptionService: SubscriptionService,
     private readonly eventStreamService: EventStreamService,
   ) {}
-
-  @Subscription(() => OnDbEventDTO, {
-    filter: (
-      payload: { onDbEvent: OnDbEventDTO },
-      variables: { input: OnDbEventInput },
-    ) => {
-      const isActionMatching =
-        !isDefined(variables.input.action) ||
-        payload.onDbEvent.action === variables.input.action;
-
-      const isObjectNameSingularMatching =
-        !isDefined(variables.input.objectNameSingular) ||
-        payload.onDbEvent.objectNameSingular ===
-          variables.input.objectNameSingular;
-
-      const isRecordIdMatching =
-        !isDefined(variables.input.recordId) ||
-        payload.onDbEvent.record.id === variables.input.recordId;
-
-      return (
-        isActionMatching && isObjectNameSingularMatching && isRecordIdMatching
-      );
-    },
-  })
-  onDbEvent(
-    @Args('input') _: OnDbEventInput,
-    @AuthWorkspace() workspace: WorkspaceEntity,
-  ) {
-    return this.subscriptionService.subscribe({
-      channel: SubscriptionChannel.DATABASE_EVENT_CHANNEL,
-      workspaceId: workspace.id,
-    });
-  }
 
   @Subscription(() => EventSubscriptionDTO, {
     nullable: true,
