+++ packages/twenty-server/src/engine/subscriptions/metadata-event/metadata-event-emitter.ts
@@ -6,8 +6,8 @@ import { isDefined } from 'twenty-shared/utils';
 
 import { getWorkspaceAuthContext } from 'src/engine/core-modules/auth/storage/workspace-auth-context.storage';
 import { type WorkspaceAuthContext } from 'src/engine/core-modules/auth/types/workspace-auth-context.type';
-import { MetadataEventBatch } from 'src/engine/metadata-event-emitter/types/metadata-event-batch.type';
-import { computeMetadataEventName } from 'src/engine/metadata-event-emitter/utils/compute-metadata-event-name.util';
+import { MetadataEventBatch } from 'src/engine/subscriptions/metadata-event/types/metadata-event-batch.type';
+import { computeMetadataEventName } from 'src/engine/subscriptions/metadata-event/utils/compute-metadata-event-name.util';
 import {
   AllMetadataEventName,
   AllMetadataEventType,
