+++ packages/twenty-server/src/engine/subscriptions/metadata-event/metadata-events-to-db.listener.ts
@@ -6,12 +6,12 @@ import { AllMetadataName } from 'twenty-shared/metadata';
 import { InjectMessageQueue } from 'src/engine/core-modules/message-queue/decorators/message-queue.decorator';
 import { MessageQueue } from 'src/engine/core-modules/message-queue/message-queue.constants';
 import { MessageQueueService } from 'src/engine/core-modules/message-queue/services/message-queue.service';
-import { type MetadataEventBatch } from 'src/engine/metadata-event-emitter/types/metadata-event-batch.type';
+import { type MetadataEventBatch } from 'src/engine/subscriptions/metadata-event/types/metadata-event-batch.type';
 import { getMetadataFlatEntityMapsKey } from 'src/engine/metadata-modules/flat-entity/utils/get-metadata-flat-entity-maps-key.util';
 import { CallWebhookJobsForMetadataJob } from 'src/engine/metadata-modules/webhook/jobs/call-webhook-jobs-for-metadata.job';
 import { WorkspaceCacheService } from 'src/engine/workspace-cache/services/workspace-cache.service';
 import { type WorkspaceCacheKeyName } from 'src/engine/workspace-cache/types/workspace-cache-key.type';
-import { WorkspaceEventEmitterService } from 'src/engine/workspace-event-emitter/workspace-event-emitter.service';
+import { MetadataEventPublisher } from 'src/engine/subscriptions/metadata-event/metadata-event-publisher';
 import { type AllMetadataEventType } from 'src/engine/workspace-manager/workspace-migration/workspace-migration-runner/types/metadata-event';
 
 @Injectable()
@@ -19,7 +19,7 @@ export class MetadataEventsToDbListener {
   constructor(
     @InjectMessageQueue(MessageQueue.webhookQueue)
     private readonly webhookQueueService: MessageQueueService,
-    private readonly workspaceEventEmitterService: WorkspaceEventEmitterService,
+    private readonly metadataEventPublisher: MetadataEventPublisher,
     private readonly workspaceCacheService: WorkspaceCacheService,
   ) {}
 
@@ -68,7 +68,7 @@ export class MetadataEventsToDbListener {
 
       const updatedCollectionHash = cacheHashes[cacheKeyName];
 
-      await this.workspaceEventEmitterService.publish({
+      await this.metadataEventPublisher.publish({
         ...metadataEventBatch,
         updatedCollectionHash,
       });
