+++ packages/twenty-server/src/engine/subscriptions/metadata-event/types/metadata-event-batch.type.ts
