+++ packages/twenty-server/src/engine/subscriptions/utils/get-channel-id-from-event-stream-id.ts
